+++ changelog.md
@@ -29,6 +29,7 @@
     - `Load layout on open` set a layout file to use when loading client, *Global preference*
     - `Use only single instance` Allow only one instance of the application to open, *Global preference*
     - `On second instance` What to do whe na second instance is opened when `Use only single instance` is enabled and no command line arguments passed, *Global preference*
+  - `#connecttime` now display last disconnect time when possible if disconnected
 - **Changed:**
   - Preferences that are now global:
     - `Spellchecking` can only now be set as a global preference and requires a restart to effect any already opening clients or windows as it can only be set on window creation